requirements.txt
alembic==1.6.5
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
colorama==0.4.4
coverage==5.5
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
packaging==21.0
pipenv==2021.5.29
platformdirs==2.0.2
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.21
toml==0.10.2
virtualenv==20.6.0
virtualenv-clone==0.5.5
Werkzeug==2.0.1
